python/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR and document processing
numpy==1.24.3
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==10.1.0

# NLP and text processing
python-dateutil==2.8.2
rapidfuzz==3.5.2
# LLM providers
groq==0.4.1
openai==1.3.5
google-generativeai==0.3.1
anthropic==0.7.7

# Data validation
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0

